refactor(push): extract handlers to sub-router factory
Move 4 push subscription handlers out of PushAPI into a dedicated
createPushRouter factory, matching the main API refactor pattern:

- lib/api/routes/push.ts — factory: createPushRouter(deps) → Router
  containing all handler implementations (subscribe, subscribeTopic,
  unsubscribeTopic, unsubscribe)
- lib/api-push.ts — trimmed from 746 to 16 lines of routing code.
  PushParameters and PushAPI shell linger; mounts the push sub-router
  via this.router.use('/', createPushRouter(...))

PushAPI remains its own Express server (port 3001) — separate failure
domain from the main API. Dead commented-out handlers (stats, vapid,
userSubscriptions, addTopics, extendSubscription, removeTopics) not
carried over.
